I've added StartActivity.
Start Activity check if preferencec contain sets of bank settings. If so than start the mainActivity. If don't than start settingsActivity.

I modify settingsActivity. If you choose none in list of bank the app is finish. If you choose The Bank, the sets of settings for him are set by default.